requirements.txt
attrs==20.3.0
aws-cdk.assets==1.92.0
aws-cdk.aws-apigateway==1.92.0
aws-cdk.aws-apigatewayv2==1.84.0
aws-cdk.aws-applicationautoscaling==1.92.0
aws-cdk.aws-autoscaling==1.84.0
aws-cdk.aws-autoscaling-common==1.92.0
aws-cdk.aws-autoscaling-hooktargets==1.84.0
aws-cdk.aws-batch==1.84.0
aws-cdk.aws-certificatemanager==1.92.0
aws-cdk.aws-cloudformation==1.92.0
aws-cdk.aws-cloudfront==1.84.0
aws-cdk.aws-cloudwatch==1.92.0
aws-cdk.aws-codebuild==1.84.0
aws-cdk.aws-codecommit==1.84.0
aws-cdk.aws-codedeploy==1.84.0
aws-cdk.aws-codeguruprofiler==1.92.0
aws-cdk.aws-codepipeline==1.84.0
aws-cdk.aws-codepipeline-actions==1.84.0
aws-cdk.aws-cognito==1.92.0
aws-cdk.aws-dynamodb==1.92.0
aws-cdk.aws-ec2==1.92.0
aws-cdk.aws-ecr==1.92.0
aws-cdk.aws-ecr-assets==1.92.0
aws-cdk.aws-ecs==1.84.0
aws-cdk.aws-efs==1.92.0
aws-cdk.aws-elasticache==1.84.0
aws-cdk.aws-elasticloadbalancing==1.84.0
aws-cdk.aws-elasticloadbalancingv2==1.92.0
aws-cdk.aws-events==1.92.0
aws-cdk.aws-events-targets==1.84.0
aws-cdk.aws-globalaccelerator==1.94.1
aws-cdk.aws-iam==1.92.0
aws-cdk.aws-kinesis==1.92.0
aws-cdk.aws-kinesisfirehose==1.84.0
aws-cdk.aws-kms==1.92.0
aws-cdk.aws-lambda==1.92.0
aws-cdk.aws-lambda-event-sources==1.92.0
aws-cdk.aws-logs==1.92.0
aws-cdk.aws-route53==1.92.0
aws-cdk.aws-route53-patterns==1.94.1
aws-cdk.aws-route53-targets==1.84.0
aws-cdk.aws-s3==1.92.0
aws-cdk.aws-s3-assets==1.92.0
aws-cdk.aws-s3-deployment==1.84.0
aws-cdk.aws-s3-notifications==1.92.0
aws-cdk.aws-sam==1.84.0
aws-cdk.aws-secretsmanager==1.84.0
aws-cdk.aws-servicecatalog==1.84.0
aws-cdk.aws-servicediscovery==1.84.0
aws-cdk.aws-signer==1.92.0
aws-cdk.aws-sns==1.92.0
aws-cdk.aws-sns-subscriptions==1.92.0
aws-cdk.aws-sqs==1.92.0
aws-cdk.aws-ssm==1.92.0
aws-cdk.aws-stepfunctions==1.84.0
aws-cdk.cloud-assembly-schema==1.92.0
aws-cdk.core==1.92.0
aws-cdk.custom-resources==1.92.0
aws-cdk.cx-api==1.92.0
aws-cdk.lambda-layer-awscli==1.84.0
aws-cdk.pipelines==1.84.0
aws-cdk.region-info==1.92.0
boto3==1.17.22
botocore==1.20.22
cattrs==1.3.0
constructs==3.2.98
elasticache-auto-discovery==1.0.0
iniconfig==1.1.1
jmespath==0.10.0
jsii==1.24.0
packaging==20.8
pluggy==0.13.1
publication==0.0.3
py==1.10.0
pymemcache==3.4.0
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
s3transfer==0.3.4
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
